introduce event with spline function and calibrate to AFG data

+++ plot_model_fits.R
@@ -23,8 +23,11 @@ library(stringr)
 library(here)
 library (Rcpp)
 library(matrixStats)
+library(splines)
+library(data.table)
+library(broom)
 
-country<- "SA"
+country<- "AFG"
 
 
 ########################################################################################
@@ -34,6 +37,7 @@ source(here("src","setup_data.R"))
 source(here("src","setup_model.R"))
 source(here("src","make_model.R"))
 source(here("src","get_sim_results.R"))
+source(here("src","spline_functions.R"))
 #Rcpp vectorized functions
 source(here("src","get_objective_vectorized.R"))
 source(here("src","goveqs_basis_rcpp.R"))
@@ -44,29 +48,35 @@ sourceCpp(here("src","compute_model_arma.cpp"))
 # source(here("src","goveqs_basis.R"))
 
 
+run_existing<-1
 
 ##################################
 
 # Load posterior samples
 posteriors <- read.table(here("output",country,"posteriors.txt"), sep = "\t" )
 theta<-posteriors%>%
-  select(A,F_risk, O_factor,imm_p)
+  select(A,F_risk, O_factor,imm_p, knot1, knot2, beta1)
 
 
+if (run_existing==0){
 ## Run model 
 
 sim<-get_sim_results(theta)
 
 saveRDS(sim, file=here("output",country,"simulations.rds"))
-
+}else{
 # ... or load previously saved sim results
 sim<-readRDS(here("output",country,"simulations.rds"))
+}
 
-  pdf(here("output",country,"model_fits.pdf")) 
+# Plot 
+
+
+  # pdf(here("output",country,"model_fits.pdf")) 
 
 plot_fits_function(sim,observations)
 
-   dev.off() 
+   # dev.off()
